+++ h2integrate/converters/ammonia/ammonia_synloop.py
@@ -2,9 +2,10 @@ import numpy as np
 import openmdao.api as om
 from attrs import field, define
 
-from h2integrate.core.utilities import BaseConfig, merge_shared_inputs
+from h2integrate.core.utilities import BaseConfig, CostModelBaseConfig, merge_shared_inputs
 from h2integrate.core.validators import gt_zero, range_val
 from h2integrate.tools.constants import H_MW, N_MW
+from h2integrate.core.model_baseclasses import CostModelBaseClass
 from h2integrate.tools.inflation.inflate import inflate_cpi, inflate_cepci
 
 
@@ -249,7 +250,7 @@ class AmmoniaSynLoopPerformanceModel(om.ExplicitComponent):
 
 
 @define
-class AmmoniaSynLoopCostConfig(BaseConfig):
+class AmmoniaSynLoopCostConfig(CostModelBaseConfig):
     """
     Configuration inputs for the ammonia synthesis loop cost model.
     *Starred inputs are from tech_config/ammonia/model_inputs/shared_parameters
@@ -299,7 +300,7 @@ class AmmoniaSynLoopCostConfig(BaseConfig):
 
     production_capacity: float = field()
     baseline_capacity: float = field()
-    base_cost_year: int = field()
+    base_cost_year: int = field(converter=int)
     capex_scaling_exponent: float = field()
     labor_scaling_exponent: float = field()
     asu_capex_base: float = field()
@@ -325,7 +326,7 @@ class AmmoniaSynLoopCostConfig(BaseConfig):
     oxygen_price_base: float = field()
 
 
-class AmmoniaSynLoopCostModel(om.ExplicitComponent):
+class AmmoniaSynLoopCostModel(CostModelBaseClass):
     """
     OpenMDAO component modeling the cost of an ammonia synthesis loop.
 
@@ -379,22 +380,24 @@ class AmmoniaSynLoopCostModel(om.ExplicitComponent):
         Annual maintenance cost
     """
 
-    def initialize(self):
-        self.options.declare("plant_config", types=dict)
-        self.options.declare("tech_config", types=dict)
-        self.options.declare("driver_config", types=dict)
-
     def setup(self):
+        target_cost_year = self.options["plant_config"]["finance_parameters"][
+            "cost_adjustment_parameters"
+        ]["target_dollar_year"]
+        self.options["tech_config"]["model_inputs"]["cost_parameters"].update(
+            {"cost_year": target_cost_year}
+        )
+
         self.config = AmmoniaSynLoopCostConfig.from_dict(
             merge_shared_inputs(self.options["tech_config"]["model_inputs"], "cost")
         )
+        super().setup()
 
         self.add_input("total_ammonia_produced", val=0.0, units="kg/year")
         self.add_input("total_hydrogen_consumed", val=0.0, units="kg/year")
         self.add_input("total_nitrogen_consumed", val=0.0, units="kg/year")
         self.add_input("total_electricity_consumed", val=0.0, units="kW*h/year")
-        self.add_output("CapEx", val=0.0, units="USD")
-        self.add_output("OpEx", val=0.0, units="USD/year")
+
         self.add_output(
             "capex_asu", val=0.0, units="USD", desc="Capital cost for air separation unit"
         )
@@ -431,13 +434,15 @@ class AmmoniaSynLoopCostModel(om.ExplicitComponent):
             "maintenance_cost", val=0.0, units="USD/year", desc="Annual maintenance cost"
         )
 
-    def compute(self, inputs, outputs):
+    def compute(self, inputs, outputs, discrete_inputs, discrete_outputs):
         ##---Scaling Ratios---
 
         # Get config values
         capacity = self.config.production_capacity  # kg NH3 / hr
         base_cap = self.config.baseline_capacity  # kg NH3 / hr
-        year = self.options["plant_config"]["plant"]["cost_year"]  # dollar year
+        year = self.options["plant_config"]["finance_parameters"]["cost_adjustment_parameters"][
+            "target_dollar_year"
+        ]  # dollar year
         base_year = self.config.base_cost_year  # dollar year
         capex_exp = self.config.capex_scaling_exponent  # unitless
         labor_exp = self.config.labor_scaling_exponent  # unitless