+++ h2integrate/converters/hydrogen/wombat_model.py
@@ -17,9 +17,11 @@ class WOMBATModelConfig(ECOElectrolyzerPerformanceModelConfig):
     """
     library_path: Path to the WOMBAT library directory, relative from this file
     if not an absolute path.
+    cost_year: dollar-year corresponding to capex value.
     """
 
     library_path: Path = field()
+    cost_year: int = field(converter=int)
 
 
 class WOMBATElectrolyzerModel(ECOElectrolyzerPerformanceModel):
@@ -42,6 +44,7 @@ class WOMBATElectrolyzerModel(ECOElectrolyzerPerformanceModel):
         self.add_output("capacity_factor", val=0.0, units=None)
         self.add_output("CapEx", val=0.0, units="USD", desc="Capital expenditure")
         self.add_output("OpEx", val=0.0, units="USD/year", desc="Operational expenditure")
+        self.add_discrete_output("cost_year", val=0, desc="Dollar year for costs")
         self.add_output(
             "percent_hydrogen_lost",
             val=0.0,
@@ -52,7 +55,7 @@ class WOMBATElectrolyzerModel(ECOElectrolyzerPerformanceModel):
             "electrolyzer_availability", val=0.0, units=None, desc="Electrolyzer availability"
         )
 
-    def compute(self, inputs, outputs):
+    def compute(self, inputs, outputs, discrete_inputs, discrete_outputs):
         super().compute(inputs, outputs)
 
         # Ensure library_path is a Path object
@@ -72,9 +75,11 @@ class WOMBATElectrolyzerModel(ECOElectrolyzerPerformanceModel):
             wombat_config["electrolyzers"]["central_electrolyzer"]["stack_capacity_kw"] * 1e-3
         )
         n_stacks = wombat_config["electrolyzers"]["central_electrolyzer"]["n_stacks"]
-        if self.config.rating != stack_capacity_mw * n_stacks:
+
+        rating_from_config = self.config.n_clusters * self.config.cluster_rating_MW
+        if rating_from_config != stack_capacity_mw * n_stacks:
             raise ValueError(
-                f"Electrolyzer rating {self.config.rating} does not match the product of "
+                f"Electrolyzer rating {rating_from_config} does not match the product of "
                 f"stack capacity {stack_capacity_mw} and number of stacks "
                 f"{n_stacks} in the WOMBAT config. "
                 "Ensure that the rating is equal to stack_capacity_kw * n_stacks."
@@ -89,7 +94,7 @@ class WOMBATElectrolyzerModel(ECOElectrolyzerPerformanceModel):
         # The "until" parameter is set to 8760 to simulate one year of operation.
         sim.run(delete_logs=True, save_metrics_inputs=False, until=8760)
 
-        scaling_factor = self.config.rating  # The baseline electrolyzer in WOMBAT is 1MW
+        scaling_factor = rating_from_config  # The baseline electrolyzer in WOMBAT is 1MW
 
         # TODO: handle cases where the project is longer than one year.
         # Do the project and divide by the project lifetime using sim.env.simulation_years
@@ -121,7 +126,7 @@ class WOMBATElectrolyzerModel(ECOElectrolyzerPerformanceModel):
         # We're currently grabbing the annual measure and since we're enforcing a single year,
         # that's fine. In the future we may need to adjust this to handle multiple years and
         # output OpEx as an array.
-        outputs["CapEx"] = self.config.electrolyzer_capex * self.config.rating * 1e3
+        outputs["CapEx"] = self.config.electrolyzer_capex * rating_from_config * 1e3
         outputs["OpEx"] = sim.metrics.opex("annual").squeeze() * scaling_factor
         outputs["electrolyzer_availability"] = sim.metrics.time_based_availability(
             "annual", "electrolyzer"
@@ -136,3 +141,5 @@ class WOMBATElectrolyzerModel(ECOElectrolyzerPerformanceModel):
         outputs["capacity_factor"] = sim.metrics.capacity_factor(
             which="net", frequency="project", by="electrolyzer"
         ).squeeze()
+
+        discrete_outputs["cost_year"] = self.config.cost_year