+++ h2integrate/core/finances.py
@@ -15,27 +15,32 @@ class AdjustedCapexOpexComp(om.ExplicitComponent):
     def setup(self):
         tech_config = self.options["tech_config"]
         plant_config = self.options["plant_config"]
-        self.discount_years = plant_config["finance_parameters"]["discount_years"]
-        self.inflation_rate = plant_config["finance_parameters"]["costing_general_inflation"]
-        self.cost_year = plant_config["plant"]["cost_year"]
+        self.inflation_rate = plant_config["finance_parameters"]["cost_adjustment_parameters"][
+            "cost_year_adjustment_inflation"
+        ]
+        self.target_dollar_year = plant_config["finance_parameters"]["cost_adjustment_parameters"][
+            "target_dollar_year"
+        ]
 
         for tech in tech_config:
             self.add_input(f"capex_{tech}", val=0.0, units="USD")
             self.add_input(f"opex_{tech}", val=0.0, units="USD/year")
+            self.add_discrete_input(f"cost_year_{tech}", val=0, desc="Dollar year for costs")
+
             self.add_output(f"capex_adjusted_{tech}", val=0.0, units="USD")
             self.add_output(f"opex_adjusted_{tech}", val=0.0, units="USD/year")
 
         self.add_output("total_capex_adjusted", val=0.0, units="USD")
         self.add_output("total_opex_adjusted", val=0.0, units="USD/year")
 
-    def compute(self, inputs, outputs):
+    def compute(self, inputs, outputs, discrete_inputs, discrete_outputs):
         total_capex_adjusted = 0.0
         total_opex_adjusted = 0.0
         for tech in self.options["tech_config"]:
             capex = float(inputs[f"capex_{tech}"][0])
             opex = float(inputs[f"opex_{tech}"][0])
-            cost_year = self.discount_years[tech]
-            periods = self.cost_year - cost_year
+            cost_year = int(discrete_inputs[f"cost_year_{tech}"])
+            periods = self.target_dollar_year - cost_year
             adjusted_capex = -npf.fv(self.inflation_rate, periods, 0.0, capex)
             adjusted_opex = -npf.fv(self.inflation_rate, periods, 0.0, opex)
             outputs[f"capex_adjusted_{tech}"] = adjusted_capex
@@ -141,15 +146,18 @@ class ProFastComp(om.ExplicitComponent):
             params.update(pf_params)
 
         check_plant_config_and_profast_params(
-            self.plant_config["plant"], params, "installation_time", "installation months"
+            self.plant_config["finance_parameters"],
+            params,
+            "installation_time",
+            "installation months",
         )
         check_plant_config_and_profast_params(
             self.plant_config["plant"], params, "plant_life", "operating life"
         )
         check_plant_config_and_profast_params(
-            self.plant_config["plant"],
+            self.plant_config["finance_parameters"],
             params,
-            "financial_analysis_start_year",
+            "analysis_start_year",
             "analysis start year",
         )
 
@@ -161,7 +169,7 @@ class ProFastComp(om.ExplicitComponent):
                 "general inflation rate"
             ]
         else:
-            gen_inflation = self.plant_config["finance_parameters"]["profast_general_inflation"]
+            gen_inflation = self.plant_config["finance_parameters"]["inflation_rate"]
 
         land_cost = 0.0
 
@@ -271,13 +279,13 @@ class ProFastComp(om.ExplicitComponent):
 
         params.setdefault(
             "installation months",
-            self.plant_config["plant"][
+            self.plant_config["finance_parameters"][
                 "installation_time"
             ],  # Add installation time to yaml default=0
         )
         params.setdefault("operating life", self.plant_config["plant"]["plant_life"])
         params.setdefault(
-            "analysis start year", self.plant_config["plant"]["financial_analysis_start_year"]
+            "analysis start year", self.plant_config["finance_parameters"]["analysis_start_year"]
         )
 
         pf = ProFAST.ProFAST()